+++ checks/executor/preludes/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "migration-executor-preludes"
+version = { workspace = true }
+edition = { workspace = true }
+license = { workspace = true }
+authors = { workspace = true }
+homepage = { workspace = true }
+publish = { workspace = true }
+rust-version = { workspace = true }
+
+[dependencies]
+serde = { workspace = true }
+anyhow = { workspace = true }
+hex = { workspace = true }
+maptos-opt-executor = { workspace = true }
+aptos-executor = { workspace = true }
+aptos-vm = { workspace = true }
+aptos-types = { workspace = true }
+thiserror = { workspace = true }
+aptos-storage-interface = { workspace = true }
+either = { workspace = true }
+migration-executor-types = { workspace = true }
+chrono = { workspace = true }
+migration-executor-test-types = { workspace = true }
+rand = { workspace = true }
+
+[lints]
+workspace = true
